Create reverseArray.java
Reverse an Array:- This is one of the simplest ways to reverse an array in Java. This algorithm iterate over the array and swap elements until you reach the midpoint. This is also known as reversing an array in place because no additional buffer is used.

In the example above, we used a temp variable to reverse the array, we swap the elements of the array. The first element is swapped with the last element. The second element id swapped with the last but one element and so on.
For instance, consider array [1, 2, 3, …., n-2, n-1, n]. We swap 1 with n, 2 with n-1, 3 with n-2 and further.